fix: using enum for keys in back and next Route components and new file to store the paths
