+++ src/ui/pages/SetPasscode/SetPasscode.tsx
@@ -10,13 +10,13 @@ import {
   SecureStorage,
   KeyStoreKeys,
 } from "../../../core/storage/secureStorage";
-import { RoutePath } from "../../../routes";
 import { PasscodeModule } from "../../components/PasscodeModule";
 import { getState } from "../../../store/reducers/stateCache";
 import { useAppDispatch, useAppSelector } from "../../../store/hooks";
 import { getNextRoute } from "../../../routes/nextRoute";
 import { updateReduxState } from "../../../store/utils";
 import { DataProps } from "../../../routes/nextRoute/nextRoute.types";
+import {RoutePath} from "../../../routes";
 
 // Based on OWASP recommendations
 const ARGON2ID_OPTIONS = {
